+++ R/write_toml.r
@@ -4,15 +4,16 @@
 #' and then writes them to a connection.
 #'
 #' @param x An R object with an [encode_toml()] method.
-#' @param con A connection object or a character string.
-#'            Passed to [base::writeLines()].
+#' @param con A connection object or a character string of the filename.
 #' @param ... Passed to [encode_toml()].
 #' @export
 write_toml <- function(x, con = stdout(), ...) {
-    if (!is_named_list(x))
-        stop("`x` must be a named list")
+    stopifnot("`x` must be a named list" = is_named_list(x))
     s <- encode_toml(x, ...)
-    writeLines(s, con = con)
+    if (is.character(con))
+        brio::writeLines(s, con = con)
+    else
+        base::writeLines(s, con = con)
 }
 
 is_named_list <- function(x) {